HOCS-1058 - Fix Migration Profile Configuration (#3)
